third_party/blink/web_tests/external/wpt/client-hints/clear-site-data/clear-site-data-cookies.https-expected.txt

This is a testharness.js-based test.
FAIL Clear-Site-Data for cookies should remove all client hints. assert_equals: expected "MissingDeviceMemory" but got "HadDeviceMemory"
Harness: the test ran to completion.
